+++ src/pages/Boards/BoardBar/BoardBar.jsx
@@ -33,7 +33,6 @@ function BoardBar() {
         justifyContent: "space-between",
         gap: 2,
         overflowX: "auto",
-        borderBottom: "1px solid white",
         bgcolor: (theme) =>
           theme.palette.mode === "dark" ? "#34495e" : "#1976d2",
         "&::-webkit-scrollbar-track": {
