--- src/pages/Boards/BoardContent/index.jsx
@@ -1,587 +1,0 @@
-import { Box, Button, Tooltip, Typography } from "@mui/material";
-import { Menu, MenuItem } from "@mui/material";
-import Divider from "@mui/material/Divider";
-import ListItemText from "@mui/material/ListItemText";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ContentCut from "@mui/icons-material/ContentCut";
-import Cloud from "@mui/icons-material/Cloud";
-import React from "react";
-import { ExpandMoreOutlined } from "@mui/icons-material";
-import DeleteForeverIcon from "@mui/icons-material/DeleteForever";
-import AddCardIcon from "@mui/icons-material/AddCard";
-import ContentPasteIcon from "@mui/icons-material/ContentPaste";
-import ContentCopyIcon from "@mui/icons-material/ContentCopy";
-import DragHandleIcon from "@mui/icons-material/DragHandle";
-import Card from "@mui/material/Card";
-import CardActions from "@mui/material/CardActions";
-import CardContent from "@mui/material/CardContent";
-import CardMedia from "@mui/material/CardMedia";
-import GroupIcon from "@mui/icons-material/Group";
-import AttachmentIcon from "@mui/icons-material/Attachment";
-import CommentIcon from "@mui/icons-material/Comment";
-
-const COLUMN_HEADER_HEIGHT = "50px";
-const COLUMN_FOOTER_HEIGHT = "50px";
-function BoardContent() {
-  const [anchorEl, setAnchorEl] = React.useState(null);
-  const open = Boolean(anchorEl);
-  const handleClick = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-  return (
-    <Box
-      sx={{
-        height: (theme) => theme.trello.boardContentHeight,
-        width: "100%",
-        display: "flex",
-        paddingX: 2,
-        justifyContent: "space-between",
-        gap: 2,
-        overflowX: "auto",
-        overflowY: "hidden",
-        borderBottom: "1px solid white",
-        bgcolor: (theme) =>
-          theme.palette.mode === "dark" ? "#34495e" : "#1976d2",
-        p: "10px 0",
-      }}
-    >
-      <Box
-        sx={{
-          bgcolor: "inherit",
-          width: "100%",
-          height: "100%",
-          display: "flex",
-          overflowX: "auto",
-          overflowY: "hidden",
-          "&::-webkit-scrollbar-track": {
-            m: 2,
-          },
-        }}
-      >
-        {/* Box Column 1 */}
-        <Box
-          sx={{
-            minWidth: "300px",
-            maxWidth: "300px",
-            bgcolor: (theme) =>
-              theme.palette.mode === "dark" ? "#333643" : "#ebecf0",
-            ml: 2,
-            borderRadius: "6px",
-            height: "fit-content",
-            maxHeight: (theme) =>
-              `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`,
-          }}
-        >
-          {/* Box Column Header*/}
-          <Box
-            sx={{
-              height: COLUMN_HEADER_HEIGHT,
-              p: 2,
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "space-between",
-            }}
-          >
-            <Typography
-              variant="h6"
-              sx={{
-                fontWeight: "bold",
-                cursor: "pointer",
-                fontSize: "1rem",
-              }}
-            >
-              Column Title
-            </Typography>
-            <Box>
-              <Tooltip title="More options">
-                <ExpandMoreOutlined
-                  sx={{ color: "text.primary", cursor: "pointer" }}
-                  id="basic-column-dropdown"
-                  aria-controls={
-                    open ? "basic-menu-column-dropdown" : undefined
-                  }
-                  aria-haspopup="true"
-                  aria-expanded={open ? "true" : undefined}
-                  onClick={handleClick}
-                />
-              </Tooltip>
-              <Menu
-                id="basic-menu-workspaces"
-                anchorEl={anchorEl}
-                open={open}
-                onClose={handleClose}
-                MenuListProps={{
-                  "aria-labelledby": "basic-column-dropdown",
-                }}
-              >
-                <MenuItem>
-                  <ListItemIcon>
-                    <AddCardIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Add new card</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCut fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Cut</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCopyIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Copy</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentPasteIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Paste</ListItemText>
-                </MenuItem>
-                <Divider />
-                <MenuItem>
-                  <ListItemIcon>
-                    <DeleteForeverIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Remote this column</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <Cloud fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Archive this column</ListItemText>
-                </MenuItem>
-              </Menu>
-            </Box>
-          </Box>
-          {/* Box Column List Card*/}
-          <Box
-            sx={{
-              p: "0 5px ",
-              m: "0 5px",
-              display: "flex",
-              flexDirection: "column",
-              gap: 1,
-              overflowX: "hidden",
-              overflowY: "auto",
-              maxHeight: (theme) =>
-                `calc(${theme.trello.boardContentHeight} - ${theme.spacing(
-                  5
-                )} - ${COLUMN_HEADER_HEIGHT} - ${COLUMN_FOOTER_HEIGHT})`,
-              "&::-webkit-scrollbar-thumb": {
-                backgroundColor: "#ced0da",
-              },
-              "&::-webkit-scrollbar-thumb:hover": {
-                backgroundColor: "#bfc2df",
-              },
-            }}
-          >
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardMedia
-                sx={{ height: 140 }}
-                image="https://scontent.fhan14-3.fna.fbcdn.net/v/t31.18172-8/23213227_664448510609643_4108551195597408808_o.jpg?stp=dst-jpg_p180x540&_nc_cat=110&ccb=1-7&_nc_sid=5f2048&_nc_eui2=AeGP72Zt421IQyI-V0F4yKZiC_s9XJlMsVcL-z1cmUyxV6FeH86dh6MFSG8W9a7RB1MpxvBV716e2c2jsNxdzw7K&_nc_ohc=KJjvAwpEa44Q7kNvgHdkhY9&_nc_ht=scontent.fhan14-3.fna&oh=00_AYBDBoe4jhMA1qn4o2-N8F5DVRnmY7N4VJz2RSklFtUYfQ&oe=6671B32F"
-                title="green iguana"
-              />
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Tuan Xuong Dev</Typography>
-              </CardContent>
-              <CardActions sx={{ p: "0 4px 8px 4px" }}>
-                <Button size="small" startIcon={<GroupIcon />}>
-                  20
-                </Button>
-                <Button size="small" startIcon={<CommentIcon />}>
-                  20
-                </Button>
-                <Button size="small" startIcon={<AttachmentIcon />}>
-                  20
-                </Button>
-              </CardActions>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-          </Box>
-          {/* Box Column Footer*/}
-          <Box
-            sx={{
-              height: COLUMN_FOOTER_HEIGHT,
-              p: 2,
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "space-between",
-            }}
-          >
-            <Button startIcon={<AddCardIcon />}>Add new card</Button>
-            <Tooltip title="Drag to move">
-              <DragHandleIcon sx={{ cursor: "pointer" }} />
-            </Tooltip>
-          </Box>
-        </Box>
-        {/* Box Column 2 */}
-        <Box
-          sx={{
-            minWidth: "300px",
-            maxWidth: "300px",
-            bgcolor: (theme) =>
-              theme.palette.mode === "dark" ? "#333643" : "#ebecf0",
-            ml: 2,
-            borderRadius: "6px",
-            height: "fit-content",
-            maxHeight: (theme) =>
-              `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)})`,
-          }}
-        >
-          {/* Box Column Header*/}
-          <Box
-            sx={{
-              height: COLUMN_HEADER_HEIGHT,
-              p: 2,
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "space-between",
-            }}
-          >
-            <Typography
-              variant="h6"
-              sx={{
-                fontWeight: "bold",
-                cursor: "pointer",
-                fontSize: "1rem",
-              }}
-            >
-              Column Title
-            </Typography>
-            <Box>
-              <Tooltip title="More options">
-                <ExpandMoreOutlined
-                  sx={{ color: "text.primary", cursor: "pointer" }}
-                  id="basic-column-dropdown"
-                  aria-controls={
-                    open ? "basic-menu-column-dropdown" : undefined
-                  }
-                  aria-haspopup="true"
-                  aria-expanded={open ? "true" : undefined}
-                  onClick={handleClick}
-                />
-              </Tooltip>
-              <Menu
-                id="basic-menu-workspaces"
-                anchorEl={anchorEl}
-                open={open}
-                onClose={handleClose}
-                MenuListProps={{
-                  "aria-labelledby": "basic-column-dropdown",
-                }}
-              >
-                <MenuItem>
-                  <ListItemIcon>
-                    <AddCardIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Add new card</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCut fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Cut</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentCopyIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Copy</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <ContentPasteIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Paste</ListItemText>
-                </MenuItem>
-                <Divider />
-                <MenuItem>
-                  <ListItemIcon>
-                    <DeleteForeverIcon fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Remote this column</ListItemText>
-                </MenuItem>
-                <MenuItem>
-                  <ListItemIcon>
-                    <Cloud fontSize="small" />
-                  </ListItemIcon>
-                  <ListItemText>Archive this column</ListItemText>
-                </MenuItem>
-              </Menu>
-            </Box>
-          </Box>
-          {/* Box Column List Card*/}
-          <Box
-            sx={{
-              p: "0 5px",
-              m: "0 5px",
-              display: "flex",
-              flexDirection: "column",
-              gap: 1,
-              overflowX: "hidden",
-              overflowY: "auto",
-              maxHeight: (theme) =>
-                `calc(${theme.trello.boardContentHeight} - ${theme.spacing(
-                  5
-                )} - ${COLUMN_HEADER_HEIGHT} - ${COLUMN_FOOTER_HEIGHT})`,
-              "&::-webkit-scrollbar-thumb": {
-                backgroundColor: "#ced0da",
-              },
-              "&::-webkit-scrollbar-thumb:hover": {
-                backgroundColor: "#bfc2df",
-              },
-            }}
-          >
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardMedia
-                sx={{ height: 140 }}
-                image="https://scontent.fhan14-3.fna.fbcdn.net/v/t31.18172-8/23213227_664448510609643_4108551195597408808_o.jpg?stp=dst-jpg_p180x540&_nc_cat=110&ccb=1-7&_nc_sid=5f2048&_nc_eui2=AeGP72Zt421IQyI-V0F4yKZiC_s9XJlMsVcL-z1cmUyxV6FeH86dh6MFSG8W9a7RB1MpxvBV716e2c2jsNxdzw7K&_nc_ohc=KJjvAwpEa44Q7kNvgHdkhY9&_nc_ht=scontent.fhan14-3.fna&oh=00_AYBDBoe4jhMA1qn4o2-N8F5DVRnmY7N4VJz2RSklFtUYfQ&oe=6671B32F"
-                title="green iguana"
-              />
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Tuan Xuong Dev</Typography>
-              </CardContent>
-              <CardActions sx={{ p: "0 4px 8px 4px" }}>
-                <Button size="small" startIcon={<GroupIcon />}>
-                  20
-                </Button>
-                <Button size="small" startIcon={<CommentIcon />}>
-                  20
-                </Button>
-                <Button size="small" startIcon={<AttachmentIcon />}>
-                  20
-                </Button>
-              </CardActions>
-            </Card>
-            <Card
-              sx={{
-                cursor: "pointer",
-                boxShadow: "0 1px 1px rgba(0, 0, 0, 0.2)",
-                overflow: "unset",
-              }}
-            >
-              <CardContent
-                sx={{
-                  p: 1.5,
-                  "&:last-child": {
-                    p: 1.5,
-                  },
-                }}
-              >
-                <Typography>Card 01</Typography>
-              </CardContent>
-            </Card>
-          </Box>
-          {/* Box Column Footer*/}
-          <Box
-            sx={{
-              height: COLUMN_FOOTER_HEIGHT,
-              p: 2,
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "space-between",
-            }}
-          >
-            <Button startIcon={<AddCardIcon />}>Add new card</Button>
-            <Tooltip title="Drag to move">
-              <DragHandleIcon sx={{ cursor: "pointer" }} />
-            </Tooltip>
-          </Box>
-        </Box>
-      </Box>
-    </Box>
-  );
-}
-
-export default BoardContent;
