+++ src/pages/Boards/_id.jsx
@@ -1,7 +1,7 @@
 import { Container } from "@mui/material";
-import AppBar from "~/components/AppBar";
-import BoardBar from "./BoardBar";
-import BoardContent from "./BoardContent";
+import AppBar from "~/components/AppBar/AppBar";
+import BoardBar from "./BoardBar/BoardBar";
+import BoardContent from "./BoardContent/BoardContent";
 
 function Board() {
   return (
